# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Urgence Santé API
  version: 1.0.0
  description: |
    Contrat REST public de l'application Urgence Santé (orientation vers les
    établissements de santé du Grand Abidjan).

    Cette spécification est la **source de vérité** partagée entre le back-end
    Spring Boot et les applications Flutter (client Dart généré). Toute
    évolution incompatible passe par une revue et une nouvelle version.
  contact:
    name: Équipe Urgence Santé
servers:
  - url: /api/v1
    description: Version 1 de l'API (préfixe de versionnement dans le chemin).
...
security: []
paths:
...
  /routes:
    get:
      tags: [Routing]
      operationId: getRoute
      summary: Calculer un itinéraire entre deux points
      parameters:
        - name: fromLat
          in: query
          required: true
          schema: { type: number, format: double, minimum: -90, maximum: 90 }
        - name: fromLon
          in: query
          required: true
          schema: { type: number, format: double, minimum: -180, maximum: 180 }
        - name: toLat
          in: query
          required: true
          schema: { type: number, format: double, minimum: -90, maximum: 90 }
        - name: toLon
          in: query
          required: true
          schema: { type: number, format: double, minimum: -180, maximum: 180 }
      responses:
        '200':
          description: Itinéraire calculé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
...
  responses:
    BadRequest:
      description: Requête invalide.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Ressource introuvable.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      description: Représentation normalisée d'une erreur (RFC 9457).
      required: [status, title]
      properties:
        type:
          type: string
          format: uri
          description: URI identifiant le type de problème.
          default: about:blank
        title:
          type: string
          description: Résumé court et lisible du type de problème.
        status:
          type: integer
          format: int32
          description: Code de statut HTTP.
        detail:
          type: string
          description: Explication spécifique à cette occurrence.
        instance:
          type: string
          format: uri
          description: URI identifiant l'occurrence du problème.
        errors:
          type: array
          description: Détail des erreurs de validation, le cas échéant.
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: Nom du champ en erreur.
        message:
          type: string
          description: Message de validation associé au champ.
...
    Route:
      type: object
      description: Itinéraire entre deux points.
      required: [distanceMeters, durationSeconds]
      properties:
        distanceMeters:
          type: number
          format: double
          description: Distance de l'itinéraire en mètres.
        durationSeconds:
          type: number
          format: double
          description: Durée estimée du trajet en secondes.